# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: HEAL+ / REDISUS Clinical API
  version: 1.0.0
  description: |
    API clínica oficial do HEAL+. A análise de imagem é um recurso de apoio à decisão,
    exige revisão por profissional habilitado e não confirma diagnóstico isoladamente.
servers:
  - url: http://localhost:5000
paths:
...
  /api/dashboard/summary:
    get:
      summary: Resumo clínico autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resumo do dashboard.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT
  responses:
    Problem:
      description: Erro HTTP no formato Problem Details (RFC 9457).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, detail, instance, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        error:
          type: string
          description: Alias temporário de code para clientes legados.
        request_id:
          type: string
...